+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -4,7 +4,7 @@ Release History
 ===============
 0.1.4 (unreleased)
 ++++++++++++++++++++
-* hostname command bug fixes (#2946, #2947, #2949)
+* Rename arg of '-n/--name' to '--hostname', and wire up default webapp name (#2946, #2947, #2949)
 
 0.1.3 (2017-04-17)
 ++++++++++++++++++++